+++ crates/app/src/commands/setup.rs
@@ -1,8 +1,9 @@
 use crate::components::create_progress_loader;
 use crate::session::MoonSession;
-use crate::systems::analyze;
 use iocraft::prelude::element;
 use moon_console::ui::{Container, Notice, StyledText, Variant};
+use moon_pdk_api::SetupToolchainInput;
+use moon_platform::PlatformManager;
 use starbase::AppResult;
 use tracing::instrument;
 
@@ -13,9 +14,24 @@ pub async fn setup(session: MoonSession) -> AppResult {
         "Downloading and installing toolchains...",
     );
 
-    let toolchain_registry = session.get_toolchain_registry().await?;
+    for platform in PlatformManager::write().list_mut() {
+        platform.setup_toolchain().await?;
+    }
 
-    analyze::load_toolchain(&toolchain_registry, &session.toolchain_config).await?;
+    session
+        .get_toolchain_registry()
+        .await?
+        .setup_toolchain_all(|registry, toolchain| SetupToolchainInput {
+            configured_version: session
+                .toolchain_config
+                .plugins
+                .get(toolchain.id.as_str())
+                .and_then(|plugin| plugin.version.clone()),
+            context: registry.create_context(),
+            toolchain_config: registry.create_config(&toolchain.id, &session.toolchain_config),
+            version: None,
+        })
+        .await?;
 
     progress.stop().await?;